test(unit): conver test for the commitizen scaffolder to vitest
for #458
